requirements.txt
requests==2.31.0
kafka-python==2.0.2
python-dotenv==1.0.0
schedule==1.2.0
pydantic==2.5.0
pydantic-settings==2.1.0
loguru==0.7.2
